+++ changelog.md
@@ -3,7 +3,8 @@ v0.9 - 01/23/2020
 - Updated Monaco Reference to v0.19.3
 - Fix broken snippet function
 - Tie css styles to CodeEditor
-- Sync CodeLanguage and Options.Language
+- Sync CodeLanguage and Options.Language, HasGlyphMargin and Options.GlyphMargin
+- Expose ReadOnly property
 
 **Breaking changes**: 
 - Use StandaloneEditorConstructionOptions for Options instead of IEditorConstructionOptions